# Avoid importing setuptools until required to speedup import