Thanks for adding the cmd_classes module. I have been using it lots of places and this seems to be working great.
This change however means that doing `import versioningit` ends up importing `setuptools` which in turn imports 
`pgk_resources` Importing pkg_resources is slow since it builds a cache of all installed packages at import time 

This tries to work around that issue in setuptools by only importing setuptools when actually invoking the cmd_classes. That way you are not paying for importing setuptools if you are using versioningit to get a runtime version


Before this change

```
In [1]: %time import versioningit
CPU times: total: 1.73 s
Wall time: 2.18 s
```

After

```
In [1]: %time import versioningit
CPU times: total: 172 ms
Wall time: 286 ms
```