@jwodder 2022-04-08
Thank you very much!  I will release a new version with this patch momentarily.